multiples updates, sales table and include datatable plugin

+++ resources/views/users/index.blade.php
@@ -17,7 +17,7 @@
 				</div>
 				<div class="panel-body">
 					<div class="table-responsive">
-						<table class="table table-hover">
+						<table class="table table-hover datatable">
 							<thead>
 								<tr>
 									<th>Id</th>
@@ -43,13 +43,13 @@
 											<a role="button" href="{{ url( 'users/'.$user->id.'/edit') }}" class="btn btn-primary">
 												<span class="glyphicon glyphicon-edit" aria-hidden="true"></span> Editar
 											</a>
-											@if($user->type != 1)
+											@if($user->type != 1 && count($user->products) == 0 ) )
 					                    	<button role="button" type="submit" class="btn btn-danger" form="form-delete-{{ $user->id }}">
 					                    		<span class="glyphicon glyphicon-remove" aria-hidden="true"></span> Eliminar
 					                    	</button>
 					                    	@endif
 										</div>
-										@if($user->type != 1)
+										@if( $user->type != 1 && count($user->products) == 0 ) )
 											{!! Form::open(['method' => 'DELETE', 'route' => ['users.destroy',$user->id], 'id' => 'form-delete-'.$user->id ]) !!} 
 	                						{!! Form::close(); !!} 
                 						@endif
